After re-initing, sync last IPNS record from network before publishing
On Heroku, the VMs have an ephemeral disk, so we recreate the IPNS
repo on each boot using a pre-configured keypair. However, this also
resets the IPNS sequence numbers to start from zero again, so it is
necessary to retrieve the latest IPNS record from the network before
publishing a new record, or a "can't replace a newer value with an
older value" error will be thrown when attempting to publish.